backend/requirements.txt
Django==5.2
djangorestframework==3.16.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.5.1
mysqlclient==2.2.7
PyYAML==6.0.2
